Split guard policies by file
